+++ src/hint_processor/builtin_hint_processor/cairo_keccak/keccak_hints.rs
@@ -1,5 +1,6 @@
 use crate::stdlib::{
     borrow::{Cow, ToOwned},
+    boxed::Box,
     collections::HashMap,
     prelude::*,
 };
@@ -93,7 +94,7 @@ pub fn compare_bytes_in_word_nondet(
     // Felt252::new(BYTES_INTO_WORD) into a lazy_static!
     let bytes_in_word = constants
         .get(BYTES_IN_WORD)
-        .ok_or(HintError::MissingConstant(BYTES_IN_WORD))?;
+        .ok_or_else(|| HintError::MissingConstant(Box::new(BYTES_IN_WORD)))?;
     let value = Felt252::new((n_bytes < bytes_in_word) as usize);
     insert_value_into_ap(vm, value)
 }
@@ -118,7 +119,7 @@ pub fn compare_keccak_full_rate_in_bytes_nondet(
     let keccak_full_rate_in_bytes = constants
         .get(KECCAK_FULL_RATE_IN_BYTES_CAIRO_KECCAK)
         .or_else(|| constants.get(KECCAK_FULL_RATE_IN_BYTES_BUILTIN_KECCAK))
-        .ok_or(HintError::MissingConstant(KECCAK_FULL_RATE_IN_BYTES))?;
+        .ok_or_else(|| HintError::MissingConstant(Box::new(KECCAK_FULL_RATE_IN_BYTES)))?;
     let value = Felt252::new((n_bytes >= keccak_full_rate_in_bytes) as usize);
     insert_value_into_ap(vm, value)
 }
@@ -152,11 +153,11 @@ pub(crate) fn block_permutation_v1(
 ) -> Result<(), HintError> {
     let keccak_state_size_felts = constants
         .get(KECCAK_STATE_SIZE_FELTS)
-        .ok_or(HintError::MissingConstant(KECCAK_STATE_SIZE_FELTS))?;
+        .ok_or_else(|| HintError::MissingConstant(Box::new(KECCAK_STATE_SIZE_FELTS)))?;
     if keccak_state_size_felts >= &Felt252::new(100_i32) {
-        return Err(HintError::InvalidKeccakStateSizeFelt252s(
+        return Err(HintError::InvalidKeccakStateSizeFelt252s(Box::new(
             keccak_state_size_felts.clone(),
-        ));
+        )));
     }
 
     let keccak_ptr = get_ptr_from_var_name("keccak_ptr", vm, ids_data, ap_tracking)?;
@@ -219,11 +220,11 @@ pub(crate) fn block_permutation_v2(
 ) -> Result<(), HintError> {
     let keccak_state_size_felts = constants
         .get(KECCAK_STATE_SIZE_FELTS)
-        .ok_or(HintError::MissingConstant(KECCAK_STATE_SIZE_FELTS))?;
+        .ok_or_else(|| HintError::MissingConstant(Box::new(KECCAK_STATE_SIZE_FELTS)))?;
     if keccak_state_size_felts >= &Felt252::from(100_i32) {
-        return Err(HintError::InvalidKeccakStateSizeFelt252s(
+        return Err(HintError::InvalidKeccakStateSizeFelt252s(Box::new(
             keccak_state_size_felts.clone(),
-        ));
+        )));
     }
 
     let keccak_ptr = get_ptr_from_var_name("keccak_ptr_start", vm, ids_data, ap_tracking)?;
@@ -256,19 +257,19 @@ fn cairo_keccak_finalize(
 ) -> Result<(), HintError> {
     let keccak_state_size_felts = constants
         .get(KECCAK_STATE_SIZE_FELTS)
-        .ok_or(HintError::MissingConstant(KECCAK_STATE_SIZE_FELTS))?;
+        .ok_or_else(|| HintError::MissingConstant(Box::new(KECCAK_STATE_SIZE_FELTS)))?;
     let block_size = constants
         .get(BLOCK_SIZE)
-        .ok_or(HintError::MissingConstant(BLOCK_SIZE))?;
+        .ok_or_else(|| HintError::MissingConstant(Box::new(BLOCK_SIZE)))?;
 
     if keccak_state_size_felts >= &Felt252::new(100_i32) {
-        return Err(HintError::InvalidKeccakStateSizeFelt252s(
+        return Err(HintError::InvalidKeccakStateSizeFelt252s(Box::new(
             keccak_state_size_felts.clone(),
-        ));
+        )));
     }
 
     if block_size >= &Felt252::new(block_size_limit) {
-        return Err(HintError::InvalidBlockSize(block_size.clone()));
+        return Err(HintError::InvalidBlockSize(Box::new(block_size.clone())));
     };
 
     let keccak_state_size_felts = keccak_state_size_felts.to_usize().unwrap();
@@ -349,10 +350,10 @@ pub(crate) fn maybe_reloc_vec_to_u64_array(
             Some(Cow::Owned(MaybeRelocatable::Int(ref num)))
             | Some(Cow::Borrowed(MaybeRelocatable::Int(ref num))) => num
                 .to_u64()
-                .ok_or_else(|| MathError::Felt252ToU64Conversion(num.clone()).into()),
-            _ => Err(VirtualMachineError::ExpectedIntAtRange(
+                .ok_or_else(|| MathError::Felt252ToU64Conversion(Box::new(num.clone())).into()),
+            _ => Err(VirtualMachineError::ExpectedIntAtRange(Box::new(
                 n.as_ref().map(|x| x.as_ref().to_owned()),
-            )),
+            ))),
         })
         .collect::<Result<Vec<u64>, VirtualMachineError>>()?;