+++ src/hint_processor/builtin_hint_processor/squash_dict_utils.rs
@@ -1,4 +1,4 @@
-use crate::stdlib::{collections::HashMap, prelude::*};
+use crate::stdlib::{boxed::Box, collections::HashMap, prelude::*};
 
 use crate::{
     hint_processor::{
@@ -34,7 +34,9 @@ fn get_access_indices(
             access_indices = Some(py_access_indices);
         }
     }
-    access_indices.ok_or_else(|| HintError::VariableNotInScopeError("access_indices".to_string()))
+    access_indices.ok_or_else(|| {
+        HintError::VariableNotInScopeError("access_indices".to_string().into_boxed_str())
+    })
 }
 
 /*Implements hint:
@@ -55,7 +57,7 @@ pub fn squash_dict_inner_first_iteration(
     //Get current_indices from access_indices
     let mut current_access_indices = access_indices
         .get(&key)
-        .ok_or_else(|| HintError::NoKeyInAccessIndices(key.clone()))?
+        .ok_or_else(|| HintError::NoKeyInAccessIndices(Box::new(key.clone())))?
         .clone();
     current_access_indices.sort();
     current_access_indices.reverse();
@@ -171,14 +173,14 @@ pub fn squash_dict_inner_used_accesses_assert(
     //Main Logic
     let access_indices_at_key = access_indices
         .get(&key)
-        .ok_or_else(|| HintError::NoKeyInAccessIndices(key.clone()))?;
+        .ok_or_else(|| HintError::NoKeyInAccessIndices(Box::new(key.clone())))?;
 
     if n_used_accesses.as_ref() != &Felt252::new(access_indices_at_key.len()) {
-        return Err(HintError::NumUsedAccessesAssertFail(
+        return Err(HintError::NumUsedAccessesAssertFail(Box::new((
             n_used_accesses.into_owned(),
             access_indices_at_key.len(),
             key,
-        ));
+        ))));
     }
     Ok(())
 }
@@ -255,22 +257,22 @@ pub fn squash_dict(
     let squash_dict_max_size = exec_scopes.get::<Felt252>("__squash_dict_max_size");
     if let Ok(max_size) = squash_dict_max_size {
         if n_accesses.as_ref() > &max_size {
-            return Err(HintError::SquashDictMaxSizeExceeded(
+            return Err(HintError::SquashDictMaxSizeExceeded(Box::new((
                 max_size,
                 n_accesses.into_owned(),
-            ));
+            ))));
         };
     };
     let n_accesses_usize = n_accesses
         .to_usize()
-        .ok_or_else(|| HintError::NAccessesTooBig(n_accesses.into_owned()))?;
+        .ok_or_else(|| HintError::NAccessesTooBig(Box::new(n_accesses.into_owned())))?;
     //A map from key to the list of indices accessing it.
     let mut access_indices = HashMap::<Felt252, Vec<Felt252>>::new();
     for i in 0..n_accesses_usize {
         let key_addr = (address + DICT_ACCESS_SIZE * i)?;
         let key = vm
             .get_integer(key_addr)
-            .map_err(|_| MemoryError::ExpectedInteger(key_addr))?;
+            .map_err(|_| MemoryError::ExpectedInteger(Box::new(key_addr)))?;
         access_indices
             .entry(key.into_owned())
             .or_default()
@@ -414,7 +416,7 @@ mod tests {
         //Execute the hint
         assert_matches!(
             run_hint!(vm, ids_data, hint_code),
-            Err(HintError::VariableNotInScopeError(x)) if x == *String::from("key")
+            Err(HintError::VariableNotInScopeError(bx)) if bx.as_ref() == "key"
         );
     }
 
@@ -654,11 +656,7 @@ mod tests {
         //Execute the hint
         assert_matches!(
             run_hint!(vm, ids_data, hint_code, &mut exec_scopes),
-            Err(HintError::NumUsedAccessesAssertFail(
-                x,
-                4,
-                y
-            )) if x == Felt252::new(5) && y == Felt252::new(5)
+            Err(HintError::NumUsedAccessesAssertFail(bx)) if *bx == (Felt252::new(5), 4, Felt252::new(5))
         );
     }
 
@@ -688,7 +686,7 @@ mod tests {
         //Execute the hint
         assert_matches!(
             run_hint!(vm, ids_data, hint_code, &mut exec_scopes),
-            Err(HintError::IdentifierNotInteger(x, y)) if x == "n_used_accesses" && y == (1,0).into()
+            Err(HintError::IdentifierNotInteger(bx)) if *bx == ("n_used_accesses".to_string(), (1,0).into())
         );
     }
 
@@ -731,7 +729,7 @@ mod tests {
         //Execute the hint
         assert_matches!(
             run_hint!(vm, HashMap::new(), hint_code),
-            Err(HintError::VariableNotInScopeError(x)) if x == *String::from("keys")
+            Err(HintError::VariableNotInScopeError(bx)) if bx.as_ref() == "keys"
         );
     }
 
@@ -964,10 +962,7 @@ mod tests {
         //Execute the hint
         assert_matches!(
             run_hint!(vm, ids_data, hint_code, &mut exec_scopes),
-            Err(HintError::SquashDictMaxSizeExceeded(
-                x,
-                y
-            )) if x == Felt252::one() && y == Felt252::new(2)
+            Err(HintError::SquashDictMaxSizeExceeded(bx)) if *bx == (Felt252::one(), Felt252::new(2))
         );
     }
 
@@ -1044,7 +1039,7 @@ mod tests {
         //Execute the hint
         assert_matches!(
             run_hint!(vm, ids_data, hint_code),
-            Err(HintError::NAccessesTooBig(x)) if x == felt_str!(
+            Err(HintError::NAccessesTooBig(bx)) if *bx == felt_str!(
                 "3618502761706184546546682988428055018603476541694452277432519575032261771265"
             )
         );