+++ src/hint_processor/cairo_1_hint_processor/dict_manager.rs
@@ -46,9 +46,9 @@ impl DictManagerExecScope {
             )
             .is_some()
         {
-            return Err(HintError::CustomHint(String::from(
-                "Segment index already in use.",
-            )));
+            return Err(HintError::CustomHint(
+                "Segment index already in use.".to_string().into_boxed_str(),
+            ));
         }
 
         Ok(dict_segment)
@@ -58,9 +58,11 @@ impl DictManagerExecScope {
     fn get_dict_tracker(&self, dict_end: Relocatable) -> Result<&DictTrackerExecScope, HintError> {
         self.trackers
             .get(&dict_end.segment_index)
-            .ok_or(HintError::CustomHint(String::from(
-                "The given value does not point to a known dictionary.",
-            )))
+            .ok_or(HintError::CustomHint(
+                "The given value does not point to a known dictionary."
+                    .to_string()
+                    .into_boxed_str(),
+            ))
     }
 
     /// Returns a mut reference for a dict tracker corresponding to a given pointer to a dict
@@ -106,22 +108,21 @@ impl DictSquashExecScope {
     /// Removes the current key, and its access indices. Should be called when only the
     /// last key access is in the corresponding indices list.
     pub fn pop_current_key(&mut self) -> Result<(), HintError> {
-        let current_key = self
-            .current_key()
-            .ok_or(HintError::CustomHint(String::from(
-                "Failed to get current key",
-            )))?;
+        let current_key = self.current_key().ok_or(HintError::CustomHint(
+            "Failed to get current key".to_string().into_boxed_str(),
+        ))?;
         let key_accesses =
             self.access_indices
                 .remove(&current_key)
-                .ok_or(HintError::CustomHint(format!(
-                    "No key accesses for key {}",
-                    current_key
-                )))?;
+                .ok_or(HintError::CustomHint(
+                    format!("No key accesses for key {current_key}").into_boxed_str(),
+                ))?;
         if !key_accesses.len().is_one() {
-            return Err(HintError::CustomHint(String::from(
-                "Key popped but not all accesses were processed.",
-            )));
+            return Err(HintError::CustomHint(
+                "Key popped but not all accesses were processed."
+                    .to_string()
+                    .into_boxed_str(),
+            ));
         }
         self.keys.pop();
         Ok(())