+++ src/hint_processor/cairo_1_hint_processor/hint_processor_utils.rs
@@ -13,12 +13,16 @@ pub(crate) fn extract_buffer(buffer: &ResOperand) -> Result<(&CellRef, Felt252),
             if let DerefOrImmediate::Immediate(val) = &bin_op.b {
                 (&bin_op.a, val.clone().value.into())
             } else {
-                return Err(HintError::CustomHint("Failed to extract buffer, expected ResOperand of BinOp type to have Inmediate b value".to_owned()));
+                return Err(HintError::CustomHint(
+                    "Failed to extract buffer, expected ResOperand of BinOp type to have Inmediate b value".to_owned().into_boxed_str()
+                ));
             }
         }
         _ => {
             return Err(HintError::CustomHint(
-                "Illegal argument for a buffer.".to_string(),
+                "Illegal argument for a buffer."
+                    .to_string()
+                    .into_boxed_str(),
             ))
         }
     };